Modified to better handle import
Previous import expected data to start and end at specific lines. Now file animation8x8.py is parsed with the only expectation that a line with date starts with '[['  so, there can be spacing between the lines or if extra code is added above or below it does not break the import.